requirements-minimal.txt
# 青藤猫教育咨询网站 - 最小依赖版本
# 当前项目实际使用的依赖包

Flask==3.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0 